requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.16
botocore==1.38.16
certifi==2025.4.26
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.9.0
jmespath==1.0.1
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
moviepy==2.1.2
numpy==2.0.2
openai==1.78.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
